+++ aladhan/__init__.py
@@ -11,16 +11,16 @@ VersionInfo = namedtuple(
 )
 
 version_info = VersionInfo(
-    major=1, minor=1, micro=0, releaselevel="final", serial=0
+    major=1, minor=2, micro=0, releaselevel="alpha", serial=0
 )
 
 __title__ = "aladhan.py"
 __author__ = "HETHAT"
-__version__ = "1.1.0"
+__version__ = "1.2.0a"
 
 
 from .client import Client
-from .base_types import *
+from .data_classes import *
 from . import methods
 from .methods import Method
 from .enums import *